requirements.txt
keyboard>=0.13.5
pyaudio>=0.2.14
numpy>=1.26.0
faster-whisper>=1.0.3
pyperclip>=1.9.0
pyautogui>=0.9.54

openai>=1.0
python-dotenv>=1.0
transformers>=4.36
torch>=2.0
sentencepiece>=0.1
